Change line-height for font variables to give more space, make PageTitleHeader, contact, privacy policy & fitness waiver pages responsive without jumping, and refactor contact page into components
