+++ src/app/(marketing)/privacy-policy/page.module.scss
@@ -3,6 +3,7 @@
 
 .body {
 	padding: 3rem 1rem;
+	transition: all 0.3s ease-in-out;
 
 	@include tablet {
 		padding: 4rem 1.5rem;
@@ -29,6 +30,7 @@
 	font-size: 1.25rem;
 	font-weight: 600;
 	line-height: 2rem;
+	transition: all 0.3s ease-in-out;
 
 	&::first-letter {
 		text-transform: capitalize;
@@ -42,17 +44,15 @@
 
 .content {
 	margin-top: 1rem;
-	font-size: 1rem;
-	line-height: 1.75rem;
+	transition: all 0.3s ease-in-out;
 
 	@include tablet {
 		margin-top: 1.5rem;
-		font-size: 1.25rem;
-		line-height: 2.25rem;
 	}
 }
 
-.link {
+a.link,
+a.link:visited {
 	text-decoration: none;
 	color: $color-red;
 }
