--- src/components/PageTitleHeader/PageTitleHeader.module.scss
@@ -1,24 +1,0 @@
-@use "styles/partials/variables" as *;
-
-.header {
-	background-color: $color-red;
-	color: $color-white;
-	padding: 7.5625rem 8.875rem 4.3125rem;
-	display: flex;
-	flex-direction: column;
-	gap: 0.5rem;
-
-	&__title {
-		font-family: "Anta";
-		font-weight: 400;
-		font-size: 4rem;
-		text-transform: uppercase;
-	}
-
-	&__subHeading {
-		font-family: "Montserrat";
-		font-weight: 400;
-		font-size: 2rem;
-		text-transform: capitalize;
-	}
-}
